build on osx too

+++ cf/897/A/897-A.c
@@ -40,7 +40,7 @@ int main(void) {
 
     fprintf(stderr, "l: %li, r: %li, c1: %c, c2: %c\n", l, r, c1, c2);
 
-    for (int i = l - 1; i <= r - 1; i++) {
+    for (long i = l - 1; i <= r - 1; i++) {
       if (s[i] == c1) {
         s[i] = c2;
       }
